% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_robust_variance}
\alias{get_robust_variance}
\title{Gets the robust variance from a linear model}
\usage{
get_robust_variance(model)
}
\arguments{
\item{model}{An lm object}
}
\value{
A list containing the robust variance and robust Std Err respectively
}
\description{
Inputs a linear model from the lm function and prints / returns the
robust variance and robust standard errors
}
\examples{
# get_robust_variance(my_linear_model)
}
