% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_studentized_residuals}
\alias{get_studentized_residuals}
\title{Gets the studentized residuals from a linear model}
\usage{
get_studentized_residuals(model)
}
\arguments{
\item{model}{An lm object}
}
\value{
Numeric vector of studentized residuals
}
\description{
Inputs a linear model from the lm function and returns the
studentized residuals as a vector
}
\examples{
# get_studentized_residuals(my_linear_model)
}
