% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lincom}
\alias{lincom}
\title{Test the hypothesis that a set of coefficients are equal to zero}
\usage{
lincom(par, mults, model, alpha = 0.05)
}
\arguments{
\item{par}{XXXX}

\item{mults}{XXX}

\item{model}{An lm object}

\item{alpha}{XXX}
}
\value{
A vector containing the test statistic and p-value
}
\description{
Inputs a linear model from the lm function and set of indices for
coefficients. Tests the hypothesis the specified set of coefficients
are equal to zero. NOTE: beta0 = 1, beta1 = 2, and so on.
}
\examples{
# Tests if
# lincom(par = c(2,4),
#        mults = c(1,7.5),
#        model=regr.a1c)
}
