% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qqnorm_studentized_residuals}
\alias{qqnorm_studentized_residuals}
\title{Creates a normal quantile-quantile plot for the studentized residuals}
\usage{
qqnorm_studentized_residuals(model)
}
\arguments{
\item{model}{An lm object}
}
\value{
A normal quantile-quantile plot for the studentized residuals
}
\description{
Inputs a linear model from the lm function and returns a
normal quantile-quantile plot for the studentized residuals
}
\examples{
# qqnorm_studentized_residuals(my_linear_model)
}
