% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{studentized_residuals_vs_fitted_plot}
\alias{studentized_residuals_vs_fitted_plot}
\title{Creates a plot of studentized residuals vs fitted values}
\usage{
studentized_residuals_vs_fitted_plot(model)
}
\arguments{
\item{model}{An lm object}
}
\value{
A plot of studentized residuals vs fitted values
}
\description{
Inputs a linear model from the lm function and creates a plot of
studentized residuals vs fitted values
}
\examples{
# studentized_residuals_vs_fitted_plot(my_linear_model)
}
