% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{subtract}
\alias{subtract}
\title{Subtracts two numbers}
\usage{
subtract(a, b)
}
\arguments{
\item{a}{numeric}

\item{b}{numeric}
}
\value{
numeric
}
\description{
Subtracts two number together using R
}
\examples{
subtract(5,8)
}
