% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{testparm}
\alias{testparm}
\title{Test the hypothesis that a set of coefficients are equal to zero}
\usage{
testparm(par, model, type = "F")
}
\arguments{
\item{par}{A vector specifying the indices of the coefficients to be tested}

\item{model}{An lm object}

\item{type}{Indicates whether an F-test or Wald test should be used}
}
\value{
A vector containing the test statistic and p-value
}
\description{
Inputs a linear model from the lm function and set of indices for
coefficients. Tests the hypothesis the specified set of coefficients
are equal to zero. NOTE: beta0 = 1, beta1 = 2, and so on.
}
\examples{
# Tests if beta1, beta4, beta5, and beta7 == 0 using an F-test
# testparm(par = c(2,5,6,8),
#          model=my_linear_model,
#          type="F)
}
