% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lincom}
\alias{lincom}
\title{Analyzing subgroup effects with a continuous modifier}
\usage{
lincom(par, mults, model, alpha = 0.05)
}
\arguments{
\item{par}{A vector specifying the indices of coefficients}

\item{mults}{A vector of multipliers}

\item{model}{An lm object}

\item{alpha}{The alpha level for the CIs and p-values}
}
\value{
A vector containing the estimate, CI, and p-value
}
\description{
Inputs a linear model from the \code{lm} function and set of indices and
multipliers for a set of coefficients. Return the estimate, CI and p-value.
NOTE: For the indices in \code{par}, beta0 = 1, beta1 = 2, and so on.
}
\examples{
# Testing beta1 + 7.5*beta3
# lincom(par = c(2,4),
#        mults = c(1,7.5),
#        model=my_linear_model)
}
